# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinPilot",
    "description": "Authenticated financial workspaces with deterministic calculations.",
    "version": "2.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Identity"
        ],
        "summary": "Create an account",
        "description": "Creates a user, a private workspace and a session cookie. With `sample_data`, the workspace holds fictional data and payment simulation is enabled. Rate limited per client address.",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "409": {
            "description": "The workspace revision changed, or the record's current state does not allow this action."
          },
          "422": {
            "description": "The request failed validation."
          },
          "429": {
            "description": "A rate limit or capacity limit was reached; retry shortly."
          }
        }
      }
    },
...
    "/api/workspace": {
      "get": {
        "tags": [
          "Workspace data"
        ],
        "summary": "Get workspace records",
        "description": "Returns the workspace records used by management forms.",
        "operationId": "workspace_api_workspace_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          }
        }
      }
    },
...
    "/api/transactions/search": {
      "post": {
        "tags": [
          "Workspace data"
        ],
        "summary": "Search transactions",
        "description": "Lists transactions newest first, filtered by account, text or category, with `limit` and `offset` paging. Filters travel in the JSON body so search text stays out of URLs and logs.",
        "operationId": "transactions_api_transactions_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionSearch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "422": {
            "description": "The request failed validation."
          }
        }
      }
    },
...
    "/api/bank/connections": {
      "get": {
        "tags": [
          "Bank connections"
        ],
        "summary": "List bank connections",
        "description": "Lists bank connections and whether bank linking is configured and available for this workspace.",
        "operationId": "connections_api_bank_connections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "503": {
            "description": "A dependency is unavailable, such as the database, a provider or its configuration."
          }
        }
      }
    },
    "/api/bank/link-token": {
      "post": {
        "tags": [
          "Bank connections"
        ],
        "summary": "Start bank linking",
        "description": "Creates a Plaid Link token. Owners and approvers only; sample workspaces cannot link banks.",
        "operationId": "link_token_api_bank_link_token_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The If-Match header does not contain a workspace revision."
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "409": {
            "description": "The workspace revision changed, or the record's current state does not allow this action."
          },
          "429": {
            "description": "A rate limit or capacity limit was reached; retry shortly."
          },
          "503": {
            "description": "A dependency is unavailable, such as the database, a provider or its configuration."
          }
        }
      }
    },
    "/api/bank/exchange": {
      "post": {
        "tags": [
          "Bank connections"
        ],
        "summary": "Finish bank linking",
        "description": "Exchanges a Plaid public token and imports supported accounts and transactions once.",
        "operationId": "exchange_api_bank_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The If-Match header does not contain a workspace revision."
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "409": {
            "description": "The workspace revision changed, or the record's current state does not allow this action."
          },
          "422": {
            "description": "The request failed validation."
          },
          "429": {
            "description": "A rate limit or capacity limit was reached; retry shortly."
          },
          "503": {
            "description": "A dependency is unavailable, such as the database, a provider or its configuration."
          }
        }
      }
    },
    "/api/bank/sync/{connection_id}": {
      "post": {
        "tags": [
          "Bank connections"
        ],
        "summary": "Sync a bank connection",
        "description": "Imports new transactions and balance snapshots for a bank connection.",
        "operationId": "sync_api_bank_sync__connection_id__post",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The If-Match header does not contain a workspace revision."
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "404": {
            "description": "The record does not exist in this workspace."
          },
          "409": {
            "description": "The workspace revision changed, or the record's current state does not allow this action."
          },
          "429": {
            "description": "A rate limit or capacity limit was reached; retry shortly."
          },
          "503": {
            "description": "A dependency is unavailable, such as the database, a provider or its configuration."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bank/connections/{connection_id}": {
      "delete": {
        "tags": [
          "Bank connections"
        ],
        "summary": "Disconnect a bank",
        "description": "Requests provider revocation and removes the stored token. Imported records are kept.",
        "operationId": "disconnect_api_bank_connections__connection_id__delete",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The If-Match header does not contain a workspace revision."
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "404": {
            "description": "The record does not exist in this workspace."
          },
          "409": {
            "description": "The workspace revision changed, or the record's current state does not allow this action."
          },
          "429": {
            "description": "A rate limit or capacity limit was reached; retry shortly."
          },
          "503": {
            "description": "A dependency is unavailable, such as the database, a provider or its configuration."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExchangeBody": {
        "properties": {
          "public_token": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Public Token"
          }
        },
        "type": "object",
        "required": [
          "public_token"
        ],
        "title": "ExchangeBody"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Registration": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "sample_data": {
            "type": "boolean",
            "title": "Sample Data",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password",
          "name"
        ],
        "title": "Registration"
      },
...
      "TransactionSearch": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "q": {
            "type": "string",
            "maxLength": 200,
            "title": "Q",
            "default": ""
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "limit": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 50
          },
          "offset": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Offset",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TransactionSearch"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}